debugging why its not writing

+++ Lab 3/myweb.cpp
@@ -199,7 +199,7 @@ void *establish_connection(void *){
                 temp = get_head(temp, &beginning, &end);
                 // printf("%s\n", temp.c_str());
                 printf("%d %d\n", beginning, written);
-                if(temp.length() < size_of_chunks){
+                if(temp.length() < (unsigned long) size_of_chunks){
                     request += "GET " + filename + " HTTP/1.1\r\nHost: " + "127.0.0.1" + "\r\n" + "Content-Range: " + to_string(temp.length()) + "-" + to_string(start + size_of_chunks) + "/" + to_string(length) + "\r\n\r\n";
                     sending_packet(socket, request);
                     temp += recieve_packets(socket);
